feat: Add StatsRing component for global CPU area chart
